#CLOUD-81: Refactoring
Task-Url: http://jira.it.dh-karlsruhe.de:8080/browse/CLOUD-81

+++ cloud/app/controllers/Dashboard.java
@@ -21,7 +21,6 @@ import views.html.playlist;
 import views.html.song;
 import views.html.user;
 
-
 public class Dashboard extends Controller {
 
 	private static final Form<Playlist> playlistform = Form
@@ -36,9 +35,9 @@ public class Dashboard extends Controller {
 	}
 
 	public static Result getPlaylist() {
-		List<Song> songs = Song.find.all();
-		List<Playlist> playlists = Playlist.find.all();
-		return ok(playlist.render(songs, playlistform, playlists,
+		List<Song> allSongs = Song.find.all();
+		List<Playlist> allPlaylists = Playlist.find.all();
+		return ok(playlist.render(allSongs, playlistform, allPlaylists,
 				new ArrayList<Song>(), null, 0l));
 	}
 
@@ -50,7 +49,7 @@ public class Dashboard extends Controller {
 	public static Result getUser() {
 		return ok(user.render());
 	}
-	
+
 	public static Result deleteEvent(long id) {
 		return TODO;
 	}
@@ -60,16 +59,14 @@ public class Dashboard extends Controller {
 		calendar.setTimeInMillis(Long.valueOf(date));
 
 		String shellCommand = "sudo date -s \"" + calendar.get(Calendar.YEAR)
-				+ "-"
-				+ (calendar.get(Calendar.MONTH) < 10 ? "0" : "")
+				+ "-" + (calendar.get(Calendar.MONTH) < 10 ? "0" : "")
 				+ (calendar.get(Calendar.MONTH) + 1) + "-"
 				+ (calendar.get(Calendar.DAY_OF_MONTH) < 10 ? "0" : "")
 				+ calendar.get(Calendar.DAY_OF_MONTH) + " "
 				+ (calendar.get(Calendar.HOUR_OF_DAY) < 10 ? "0" : "")
-				+ calendar.get(Calendar.HOUR_OF_DAY) + ":" 
+				+ calendar.get(Calendar.HOUR_OF_DAY) + ":"
 				+ (calendar.get(Calendar.MINUTE) < 10 ? "0" : "")
-				+ calendar.get(Calendar.MINUTE) + ":"
-				+ "00\"";
+				+ calendar.get(Calendar.MINUTE) + ":" + "00\"";
 		ShellCommand command = new ShellCommand(shellCommand);
 		command.executeShellCommand();
 
@@ -78,71 +75,44 @@ public class Dashboard extends Controller {
 
 	public static Result editPlaylist(long id) {
 
-		Playlist foundPlaylist = Playlist.find.byId(id);
-		foundPlaylist.updateListOfSongs();
-		Form<Playlist> filledPlaylist = playlistform.fill(foundPlaylist);
-
-		List<Song> songs = Song.find.all();
-		songs = Playlist.deleteSameSongs(songs, foundPlaylist.songListe);
-
+		List<Song> allSongs = Song.find.all();
 		List<Playlist> allPlaylists = Playlist.find.all();
 
-		return ok(playlist.render(songs, filledPlaylist, allPlaylists,
-				foundPlaylist.songListe, foundPlaylist.name, id));
+		Playlist playlist = Playlist.find.byId(id);
+		playlist.stringToList();
+		Form<Playlist> filledPlaylist = playlistform.fill(playlist);
+
+		allSongs = Playlist.deleteSameSongs(playlist.getSongList());
+
+		return ok(views.html.playlist.render(allSongs, filledPlaylist,
+				allPlaylists, playlist.getSongList(), playlist.name, id));
 	}
 
 	public static Result deletePlaylist(long id) {
 
 		Playlist playlist = Playlist.find.byId(id);
-		playlist.delete();
+		if (playlist != null) {
+			playlist.delete();
+		}
 
 		return redirect(routes.Dashboard.getPlaylist());
 	}
 
-	public static Result savePlaylist(long idToDelete) {
+	public static Result savePlaylist(long id) {
 
 		Form<Playlist> formPlaylist = Form.form(Playlist.class)
 				.bindFromRequest();
 		Map<String, String> myPlaylist = formPlaylist.data();
 
-		String name = myPlaylist.remove("name");
-		Iterator<String> it = myPlaylist.values().iterator();
+		String playlistName = myPlaylist.remove("name");
+		Iterator<String> iterator = myPlaylist.values().iterator();
 
+		// No songs added
 		if (myPlaylist.size() < 1) {
-			// No songs added
 			return redirect(routes.Dashboard.getPlaylist());
 		}
-
-		if (idToDelete > 0) {
-			Playlist.find.byId(idToDelete).delete();
-		}
-
-		ArrayList<Song> songs = new ArrayList<Song>();
-
-		long duration = calculateDuration(it);
-
-		while (it.hasNext()) {
-			long id = Integer.valueOf((String) it.next());
-			Song song = Song.find.byId(id);
-			songs.add(song);
-		}
-
-		Playlist playlist = new Playlist(name, songs, duration);
-		playlist.save();
-
+		Playlist.savePlaylist(id, playlistName, iterator);
 		return redirect(routes.Dashboard.getPlaylist());
-	}
-
-	public static long calculateDuration(Iterator<String> it) {
-		long duration = 0;
-		long id;
-		while (it.hasNext()) {
-			id = Integer.valueOf((String) it.next());
-
-			Song song = Song.find.byId(id);
-			duration += song.duration;
-		}
-		return duration;
 	}
 
 	public static Result uploadSong() {
